runtime/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.24
motor>=3.3
python-dotenv>=1.0
langgraph>=0.2.60
langchain-core>=0.3
sentence-transformers>=3.0
numpy>=1.26
